Fixed company folder inception
