--- companies/companies/admin.py
@@ -1,5 +1,0 @@
-from django.contrib import admin
-from companies.models import Company
-# Register your models here.
-
-admin.site.register(Company)
