--- companies/companies/templates/companies.html.j2
@@ -1,108 +1,0 @@
-{% extends 'global/base.html.j2' %}
-
-{% block content %}
-<div class="my-3 my-md-5">
-    <div class="container">
-        <div class="card">
-            <div class="card-header">
-                <h3 class="card-title">Companies</h3>
-                
-                    {% if user.is_authenticated %}
-                <div class="col text-right">
-                    <a class="button" href="{{ url('companies:new') }}">
-                        <button type="submit" class="btn btn-primary ml-auto">Add Company</button>
-                    </a>
-                </div>
-                
-                    {% endif %}
-                <div class="col-lg-3 ml-auto">
-                    <div class="row gutters-xs">
-                        <div class="col">
-                            <input class="form-control" placeholder="Search for..." type="text">
-                        </div>
-                        <span class="col-auto">
-                            <button class="btn btn-secondary" type="button"><i class="fe fe-search"></i></button>
-                        </span>
-                    </div>
-                </div>
-            </div>
-            <div class="table-responsive">
-                <table class="table card-table table-vcenter text-nowrap">
-                    <thead>
-                        <tr>
-                            <th class="w-1">Checked</th>
-                            <th>Name</th>
-                            <th>Industries</th>
-                            <th>Size</th>
-                            <th>Updated</th>
-                            <th>Status</th>
-                            <th>Donated</th>
-                            <th></th>
-                            <th></th>
-                        </tr>
-                    </thead>
-                    <tbody>
-                        {% for company in companies %}
-                        <tr>
-                            <td>
-                                <div>
-                                    <label class="custom-control custom-checkbox custom-control-inline">
-                                        <input class="custom-control-input" name="example-inline-checkbox1" value="option1"
-                                            checked="" type="checkbox">
-                                        <span class="custom-control-label"></span>
-                                    </label>
-                                </div>
-                            </td>
-                            <td><a href="" class="text-inherit">{{ company.name }}</a></td>
-                            <td>
-                                {% for industry in company.industries.all() %}
-                                <span class="tag tag-{{ industry.color }}">{{ industry.name }}</span>
-                                {% endfor %}
-                            </td>
-                            <td>
-                                {%if company.size == 'S'%}
-                                <span class="tag tag-green">Small</span>
-                                {% endif %}
-                                {%if company.size == 'M'%}
-                                <span class="tag tag-yellow">Medium</span>
-                                {% endif %}
-                                {%if company.size == 'L'%}
-                                <span class="tag tag-red">Large</span>
-                                {% endif %}
-                            </td>
-                            <td> {{ company.updated }} </td>
-                            <td>
-                                {%if company.status == 'U'%}
-                                <span class="status-icon bg-secondary"></span> Uncontacted
-                                {% endif %}
-                                {%if company.status == 'C'%}
-                                <span class="status-icon bg-warning"></span> Contacted
-                                {% endif %}
-                                {%if company.status == 'D'%}
-                                <span class="status-icon bg-success"></span> Donated
-                                {% endif %}
-
-                            </td>
-                            <td>${{ company.donated }}</td>
-                            <td class="text-right">
-                            </td>
-                            <td>
-                                {% if user.is_authenticated %}
-                                <a class="icon" href="{{ url('companies:edit', pk=company.pk) }}">
-                                    <i class="fe fe-edit"></i>
-                                </a>
-                                {% endif %}
-                            </td>
-                        </tr>
-                        {% endfor %}
-                    </tbody>
-                </table>
-            </div>
-        </div>
-    </div>
-</div>
-{% endblock %}
-
-{% block js %}
-{# import javascript here if you want it otherwise delete this block #}
-{% endblock %}
